Moved the GameActionHandler in the game.action package
