+++ console-client/src/main/java/machinestrike/client/console/action/ClientActionHandler.java
@@ -1,6 +1,6 @@
 package machinestrike.client.console.action;
 
-import machinestrike.action.GameActionHandler;
+import machinestrike.game.action.GameActionHandler;
 import org.jetbrains.annotations.NotNull;
 
 public interface ClientActionHandler extends GameActionHandler {
